--- test/libgetpid.h
@@ -1,6 +1,0 @@
-#ifndef __LIBGETPID_H__
-#define __LIBGETPID_H__
-
-extern void getpid_func();
-
-#endif
