--- test/librandom.h
@@ -1,6 +1,0 @@
-#ifndef __LIBRANDOM_H__
-#define __LIBRANDOM_H__
-
-extern void func();
-
-#endif
