+++ test/snippets/getpid.c
@@ -1,0 +1,2 @@
+pid_t pid = getpid();
+printf("[%s] getpid() yielded %d\n", where, pid);
