+++ test/snippets/include_headers.h
@@ -1,0 +1,4 @@
+#include <stdio.h>
+#include <sys/types.h>
+#include <unistd.h>
+#include <sys/random.h>
