--- test/use_lib_random.c
@@ -1,6 +1,0 @@
-#include "librandom.h"
-
-int main() {
-  func();
-  return 0;
-}
